+++ bika/health/catalog/__init__.py
@@ -1,0 +1,14 @@
+# This file is part of Bika Health
+#
+# Copyright 2011-2016 by it's authors.
+# Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+
+# Catalog IDs static constant
+from .patient_catalog import CATALOG_PATIENT_LISTING
+# Catalog classes
+from .patient_catalog import BikaPatientCatalog
+from .patient_catalog import BikaHealthCatalogPatientListing
+# Catalog public functions
+from .catalog_utilities import getCatalogDefinitions
+from .catalog_utilities import getCatalogExtensions
+from .catalog_utilities import getCatalog
